Tests/LibWeb/Text/input/missing-font-family-serialization.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
  @font-face { src: local('Test'); }
</style>
<script>
  test(() => {
        let rule = document.querySelector("style").sheet.cssRules[0];
        println(rule.cssText);
    });
</script>
