Tests/LibWeb/Text/expected/missing-font-family-serialization.txt

@font-face { src: local("Test"); }
